project_rag/requirements.txt
streamlit>=1.35.0
sentence-transformers>=2.7.0
groq>=0.9.0
requests>=2.31.0
numpy>=1.26.0
python-dotenv>=1.0.0
msgpack>=1.0.7
pypdf>=4.0.0
reportlab>=4.1.0